cpu: fix msvx15 narrowing conversion compilation error

+++ src/cpu/x64/rnn/jit_diff_weights_peephole.cpp
@@ -32,8 +32,9 @@ jit_diff_weights_peephole_t::jit_diff_weights_peephole_t(
     , tail_size_(dhc_block_size % simd_w_)
     , io_(this, mayiuse(avx512_core_bf16) ? avx512_core_bf16 : avx512_core,
               {c_states_dt_, scratch_dt_, dst_dt_}, {},
-              io::io_tail_conf_t {simd_w_, static_cast<std::size_t>(tail_size_),
-                      tail_opmask_, 0u, reg_tmp_}) {}
+              io::io_tail_conf_t {static_cast<std::size_t>(simd_w_),
+                      static_cast<std::size_t>(tail_size_), tail_opmask_, 0,
+                      reg_tmp_}) {}
 
 void jit_diff_weights_peephole_t::generate() {
     preamble();
